+++ packages/yourloops/components/patient/patient-row.tsx
@@ -52,6 +52,8 @@ import { useHistory } from 'react-router-dom'
 import { Patient } from '../../lib/data/patient'
 import RemovePatientDialog from './remove-patient-dialog'
 import EmailOpenIcon from '../icons/EmailOpenIcon'
+import RemoveDirectShareDialog from '../dialogs/remove-direct-share-dialog'
+import { usePatientContext } from '../../lib/patient/provider'
 
 const patientListStyle = makeStyles(
   (theme: Theme) => {
@@ -96,6 +98,7 @@ const patientListStyle = makeStyles(
 
 const PatientRow: FunctionComponent<PatientRowProps> = ({ patient, filter }) => {
   const historyHook = useHistory()
+  const patientHook = usePatientContext()
   const { t } = useTranslation('yourloops')
   const classes = patientListStyle()
   const patientListCommonClasses = patientListCommonStyle()
@@ -111,11 +114,20 @@ const PatientRow: FunctionComponent<PatientRowProps> = ({ patient, filter }) => 
   const isAlreadyInATeam = PatientUtils.isInAtLeastATeam(patient)
   const hasUnreadMessages = patient.metadata.unreadMessagesSent > 0
 
+  const removeDirectShareDialogProps = {
+    userToRemove: {
+      id: userId,
+      fullName: patient.profile.fullName,
+      email: patient.profile.email
+    }
+  }
+
   const {
     computeRowInformation,
     flagPatient,
     trNA,
     isUserHcp,
+    isUserCaregiver,
     isFlagged
   } = usePatientRow({ patient, classes })
 
@@ -147,6 +159,14 @@ const PatientRow: FunctionComponent<PatientRowProps> = ({ patient, filter }) => 
 
   const onCloseRemovePatientDialog = (): void => {
     setPatientToRemove(null)
+  }
+
+  const onCloseRemoveDirectShareDialog = (shouldRefresh?: boolean): void => {
+    setPatientToRemove(null)
+
+    if (shouldRefresh) {
+      patientHook.refresh()
+    }
   }
 
   const { lastUpload } = useMemo(() => getMedicalValues(medicalData, trNA), [medicalData, trNA])
@@ -276,6 +296,13 @@ const PatientRow: FunctionComponent<PatientRowProps> = ({ patient, filter }) => 
         <RemovePatientDialog
           patient={patient}
           onClose={onCloseRemovePatientDialog}
+        />
+      }
+
+      {patientToRemove && isUserCaregiver &&
+        <RemoveDirectShareDialog
+          userToRemove={removeDirectShareDialogProps.userToRemove}
+          onClose={onCloseRemoveDirectShareDialog}
         />
       }
     </React.Fragment>